requirements.txt

viam-sdk==0.44.1

typing-extensions

adafruit-circuitpython-hcsr04
adafruit-blinka
